duplicates removing v1
