+++ utility.py
@@ -1,3 +1,5 @@
+from classes.sequence import ProteinSequence
+from classes.antiboby import Antibody
 from sys import stdout
 from datetime import datetime
 from os import makedirs, path
@@ -5,6 +7,94 @@ import matplotlib.pyplot as plt
 from pandas import read_csv
 from matplotlib.ticker import FuncFormatter
 import numpy as np
+import gc
+
+
+# parsing of data file with sequences
+def sequences_parser(filename: str)->list:
+
+    antibodies = []
+    str = ''
+
+    with open(filename) as file:
+        lines = file.readlines()
+
+        # smart bar for getting the time remaining
+        progress_label = 'Parsing the input data'
+        show_progress(progress_label, 0.0)
+        index = 0
+
+        for line in lines:
+
+            # for appropriate work of smart bar
+            index += 1
+            show_progress(progress_label, float(index) / float(len(lines)))
+
+            line = line.rstrip()
+            # the way to find new seq and initiate new object
+            if line[0] == '>':
+
+                # add the sequence to its antibody object as to a previous object in antibodies list
+                if len(antibodies) != 0 and str != '':
+                    protein_sequence = ProteinSequence(seq=str)
+                    antibodies[-1].protein_sequence = protein_sequence
+                    str=''
+
+                # generate hot points in class initiating
+                laa = line.split('|')
+                name = laa[0][1:]
+                host = laa[2]
+                res = laa[3]
+                # initiate an object of class Antibody
+                # todo: write a filter to rid out of the sequence duplicates
+                antibody = Antibody(name=name, host=host, resource_of_origin=res)
+
+                # define what kind of seq do we have
+                if laa[5] == 'protein':
+                    antibody.protein_sequence = True
+                    antibody.nucleotide_sequence = False
+                elif laa[5] == 'nucleotide':
+                    antibody.protein_sequence = False
+                    antibody.nucleotide_sequence = True
+                antibodies.append(antibody)
+
+            else:
+                str += line
+                # to fill the last line
+                if line == lines[-1].rstrip():
+                    protein_sequence = ProteinSequence(seq=str)
+                    antibodies[-1].protein_sequence = protein_sequence
+
+    return antibodies
+
+
+# remove repeats from list of sequences
+def remove_repeats(antibodies: list) -> list:
+    new_antibodies_list = []
+
+    # smart bar for getting the time remaining
+    progress_label = 'Removing sequence duplicates'
+    show_progress(progress_label, 0.0)
+    index = 0
+
+    # left only antibodies with unique dictionaries
+    for antibody in antibodies:
+
+        # for appropriate work of smart bar
+        index += 1
+        show_progress(progress_label, float(index) / float(len(antibodies)))
+
+        if len(new_antibodies_list) != 0:
+            for new_antibody in new_antibodies_list:
+                if antibody.protein_sequence.whole_seq != new_antibody.protein_sequence.whole_seq:
+                    new_antibodies_list.append(antibody)
+        else:
+            new_antibodies_list.append(antibody)
+
+        gc.collect()
+
+    print(len(new_antibodies_list))
+    return new_antibodies_list
 
 
 # smart progress bar
